+++ Sources/KeyboardKit/_Deprecated/EnglishSecondaryCalloutActionProvider.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "EnglishCalloutActionProvider")
+public typealias EnglishSecondaryCalloutActionProvider = EnglishCalloutActionProvider