+++ Sources/KeyboardKit/_Deprecated/LocalizedSecondaryCalloutActionProvider.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "LocalizedCalloutActionProvider")
+public typealias LocalizedSecondaryCalloutActionProvider = LocalizedCalloutActionProvider