+++ Sources/KeyboardKit/_Deprecated/StandardSecondaryCalloutActionProvider.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "StandardCalloutActionProvider")
+public typealias StandardSecondaryCalloutActionProvider = StandardCalloutActionProvider