+++ Tests/KeyboardKitTests/Callouts/StandardCalloutActionProviderTests.swift
@@ -1,5 +1,5 @@
 //
-//  StandardSecondaryCalloutActionProviderTests.swift
+//  StandardCalloutActionProviderTests.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-01-06.
@@ -11,16 +11,16 @@ import Nimble
 import Foundation
 import KeyboardKit
 
-class StandardSecondaryCalloutActionProviderTests: QuickSpec {
+class StandardCalloutActionProviderTests: QuickSpec {
     
     override func spec() {
         
-        var provider: StandardSecondaryCalloutActionProvider!
+        var provider: StandardCalloutActionProvider!
         var context: KeyboardContext!
         
         beforeEach {
             context = KeyboardContext(controller: MockKeyboardInputViewController())
-            provider = StandardSecondaryCalloutActionProvider(context: context)
+            provider = StandardCalloutActionProvider(context: context)
         }
         
         describe("localized providers") {
@@ -32,9 +32,9 @@ class StandardSecondaryCalloutActionProviderTests: QuickSpec {
             }
             
             it("accepts custom providers") {
-                provider = StandardSecondaryCalloutActionProvider(
+                provider = StandardCalloutActionProvider(
                     context: context,
-                    providers: [StandardSecondaryCalloutActionProvider.standardProvider])
+                    providers: [StandardCalloutActionProvider.standardProvider])
                 let providers = provider.providerDictionary.dictionary
                 expect(providers.keys.count).to(equal(1))
                 expect(providers[KeyboardLocale.english.id] is EnglishSecondaryCalloutActionProvider).to(beTrue())
@@ -46,7 +46,7 @@ class StandardSecondaryCalloutActionProviderTests: QuickSpec {
             it("supports English") {
                 context.locale = Locale(identifier: KeyboardLocale.english.id)
                 let action = KeyboardAction.character("a")
-                let actions = provider.secondaryCalloutActions(for: action)
+                let actions = provider.calloutActions(for: action)
                 let expected = "aàáâäæãåā".map { KeyboardAction.character(String($0)) }
                 expect(actions).to(equal(expected))
             }
@@ -54,7 +54,7 @@ class StandardSecondaryCalloutActionProviderTests: QuickSpec {
             it("has fallback support for specific locale") {
                 context.locale = Locale(identifier: "en-US")
                 let action = KeyboardAction.character("a")
-                let actions = provider.secondaryCalloutActions(for: action)
+                let actions = provider.calloutActions(for: action)
                 let expected = "aàáâäæãåā".map { KeyboardAction.character(String($0)) }
                 expect(actions).to(equal(expected))
             }
@@ -62,7 +62,7 @@ class StandardSecondaryCalloutActionProviderTests: QuickSpec {
             it("has fallback support for non-supported locale") {
                 context.locale = Locale(identifier: KeyboardLocale.german.id)
                 let action = KeyboardAction.character("a")
-                let actions = provider.secondaryCalloutActions(for: action)
+                let actions = provider.calloutActions(for: action)
                 let expected = "aàáâäæãåā".map { KeyboardAction.character(String($0)) }
                 expect(actions).to(equal(expected))
             }