+++ Tests/KeyboardKitTests/Keyboard/KeyboardInputViewControllerTests.swift
@@ -141,13 +141,13 @@ class KeyboardInputViewControllerTests: QuickSpec {
             
             it("has standard instances by default") {
                 expect(vc.autocompleteProvider as? DisabledAutocompleteProvider).toNot(beNil())
+                expect(vc.calloutActionProvider as? StandardCalloutActionProvider).toNot(beNil())
                 expect(vc.keyboardActionHandler as? StandardKeyboardActionHandler).toNot(beNil())
                 expect(vc.keyboardAppearance as? StandardKeyboardAppearance).toNot(beNil())
                 expect(vc.keyboardBehavior as? StandardKeyboardBehavior).toNot(beNil())
                 expect(vc.keyboardFeedbackHandler as? StandardKeyboardFeedbackHandler).toNot(beNil())
                 expect(vc.keyboardInputSetProvider as? StandardKeyboardInputSetProvider).toNot(beNil())
                 expect(vc.keyboardLayoutProvider as? StandardKeyboardLayoutProvider).toNot(beNil())
-                expect(vc.keyboardSecondaryCalloutActionProvider as? StandardSecondaryCalloutActionProvider).toNot(beNil())
             }
         }
         
@@ -157,7 +157,7 @@ class KeyboardInputViewControllerTests: QuickSpec {
                 let actionContext = vc.actionCalloutContext
                 let layoutProvider = vc.keyboardLayoutProvider as? StandardKeyboardLayoutProvider
                 expect(layoutProvider?.inputSetProvider).to(be(vc.keyboardInputSetProvider))
-                expect(actionContext.actionProvider).to(be(vc.keyboardSecondaryCalloutActionProvider))
+                expect(actionContext.actionProvider).to(be(vc.calloutActionProvider))
                 expect(actionContext.actionHandler).to(be(vc.keyboardActionHandler))
             }
             
@@ -172,7 +172,7 @@ class KeyboardInputViewControllerTests: QuickSpec {
             }
             
             it("is done for keyboard secondary callout action provider") {
-                vc.keyboardSecondaryCalloutActionProvider = StandardSecondaryCalloutActionProvider(context: .preview)
+                vc.calloutActionProvider = StandardCalloutActionProvider(context: .preview)
                 verifyRefresh()
             }
         }